requirements.txt
# Core dependencies only
fastapi==0.104.1
uvicorn==0.24.0

# LangChain essentials
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2

# Vector store and embeddings
chromadb==0.4.18
sentence-transformers==2.7.0

# Document processing
PyPDF2==3.0.1
python-multipart==0.0.6

# HTTP client for tools
httpx==0.25.2

# Configuration
python-dotenv==1.0.0

# RAG Evaluation
ragas==0.3.4